+++ tests/Unit/Private/Get-DscProperty.Tests.ps1
@@ -752,4 +752,66 @@ $script:mockResourceBaseInstance.MyResourceProperty2 = 'MockValue5'
             }
         }
     }
+
+    Context 'When excluding specific property names' {
+        BeforeAll {
+            <#
+                Must use a here-string because we need to pass 'using' which must be
+                first in a scriptblock, but if it is outside the here-string then
+                PowerShell will fail to parse the test script.
+            #>
+            $inModuleScopeScriptBlock = @'
+class MyMockResource
+{
+[DscProperty(Key)]
+[System.String]
+$MyResourceKeyProperty1
+
+[DscProperty(Key)]
+[System.String]
+$MyResourceKeyProperty2
+
+[DscProperty(Mandatory)]
+[System.String]
+$MyResourceMandatoryProperty
+
+[DscProperty()]
+[System.String]
+$MyResourceProperty1
+
+[DscProperty()]
+[System.String]
+$MyResourceProperty2
+
+[DscProperty(NotConfigurable)]
+[System.String]
+$MyResourceReadProperty
+}
+
+$script:mockResourceBaseInstance = [MyMockResource]::new()
+$script:mockResourceBaseInstance.MyResourceKeyProperty1 = 'MockValue1'
+$script:mockResourceBaseInstance.MyResourceKeyProperty2 = 'MockValue2'
+$script:mockResourceBaseInstance.MyResourceMandatoryProperty = 'MockValue3'
+$script:mockResourceBaseInstance.MyResourceProperty1 = 'MockValue5'
+$script:mockResourceBaseInstance.MyResourceProperty2 = 'MockValue6'
+'@
+
+            InModuleScope -ScriptBlock ([Scriptblock]::Create($inModuleScopeScriptBlock))
+        }
+
+        It 'Should return the correct value' {
+            InModuleScope -ScriptBlock {
+                $result = Get-DscProperty -ExcludeName @('MyResourceKeyProperty1', 'MyResourceProperty1') -HasValue -InputObject $script:mockResourceBaseInstance
+
+                $result | Should -BeOfType [System.Collections.Hashtable]
+
+                $result.Keys | Should -Not -Contain 'MyResourceKeyProperty1' -Because 'the property was excluded'
+                $result.Keys | Should -Not -Contain 'MyResourceProperty1' -Because 'the property was excluded'
+
+                $result.Keys | Should -Contain 'MyResourceKeyProperty2' -Because 'the property has a non-null value and was not excluded'
+                $result.Keys | Should -Contain 'MyResourceProperty2' -Because 'the property has a non-null value and was not excluded'
+                $result.Keys | Should -Contain 'MyResourceMandatoryProperty' -Because 'the property has a non-null value and was not excluded'
+            }
+        }
+    }
 }