+++ CHANGELOG.md
@@ -95,6 +95,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
       the PowerShell SqlServer stub module when a test has run.
 - SqlWindowsFirewall
   - Added integration tests for SqlWindowsFirewall ([issue #747](https://github.com/dsccommunity/SqlServerDsc/issues/747)).
+- `Get-DscProperty`
+  - Added parameter `ExcludeName` to exclude property names from being returned.
 
 ### Changed
 
@@ -307,6 +309,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     and correct localized string ID for each string.
 - `Set-SqlDscDatabasePermission`
   - Minor code cleanup.
+- `ConvertTo-Reason`
+  - Fix to handle `$null` values on Windows PowerShell.
 
 ## [15.2.0] - 2021-09-01